% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to a server}
\usage{
connect(
  server = getOption("stefan_server"),
  user = getOption("stefan_user"),
  verbose = TRUE
)
}
\arguments{
\item{server}{A remote server to connect to.}

\item{user}{User name to authenticate with the server.}

\item{verbose}{Whether to show output from the \code{ssh} package. Defaults to
\code{FALSE}.}
}
\description{
Persistently connect to a server. Stores the session object in the cache
so that you don't have to re-authenticate with every call to
\code{\link{remotely}}.
}
\examples{
\dontrun{
# connect once
connect()

# no need to connect every time
remotely(print("hello"))
}
}
