% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotely.R
\name{remotely}
\alias{remotely}
\title{Run code on a remote server}
\usage{
remotely(
  expr,
  rmt_path = random_name(),
  pkgs = NULL,
  objs = NULL,
  env = parent.frame(),
  session = active_session(),
  server = getOption("stefan_server"),
  user = getOption("stefan_user"),
  allow_install = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{expr}{R expression with a return value, i.e. the last line should
return a meaningful value}

\item{rmt_path}{Directory on the remote server where temporary data are
stored. The path will be removed after the function exits.}

\item{pkgs}{Additional packages to attach. If any package passed to this
argument is not installed on the remote system, will try to install them
using \code{pak::pkg_install}. This may fail for a couple reasons. Therefore,
it is recommended to manually install the packages using ``.}

\item{objs}{Further objects not included in \code{env} that should be attached
to the remote session. This could be specific objects or non-compiled
functions from a package or an object from a non-parent environment.}

\item{env}{Environment in which \code{expr} should be evaluated. The environment
object is transferred to the remote server as an \code{.rds} file. Note that
package environments and other "named" parent environments (e.g. tools from
RStudio or renv) are omitted in the process. Unnamed parent environments
and the global environment are assembled in a list of environments and then
unwrapped and merged together in the remote server. Defaults to the
caller environment.}

\item{server}{A remote server to connect to.}

\item{user}{User name to authenticate with the server.}

\item{verbose}{Whether to show output from the \code{ssh} package. Defaults to
\code{FALSE}.}
}
\value{
The result of \code{expr}.
}
\description{
Run code on a remote server
}
