% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{rmt_installed}
\alias{rmt_installed}
\title{Shorthands}
\usage{
rmt_installed(
  rmt_path = random_name(),
  session = active_session(),
  server = getOption("stefan_server"),
  user = getOption("stefan_user")
)
}
\arguments{
\item{rmt_path}{Directory on the remote server where temporary data are
stored. The path will be removed after the function exits.}

\item{server}{A remote server to connect to.}

\item{user}{User name to authenticate with the server.}

\item{pkg}{A package to install. Supports the syntax of
\code{\link[pak]{pkg_install}}}

\item{...}{Further arguments passed to \code{\link[pak]{pkg_install}}.}
}
\description{
A couple of shorthand functions that do common chores remotely.

\itemize{
\item{\code{rmt_installed} lists all installed packages}
\item{\code{rmt_install} installs a package}
\item{\code{rmt_session_info} gives \code{sessionInfo()} remotely}
}
}
